feat: artist Lesson style 태그 규칙 추가 및 개선

+++ src/main/java/com/muzin/mu/zin/service/artistProfile/ArtistStyleService.java
@@ -45,23 +45,28 @@ public class ArtistStyleService {
 
     // 스타일 태그 수정 및 생성
     @Transactional
-    public ApiRespDto<?> setMyStyleTags(PrincipalUser principal, ArtistStyleSetRequest req) {
+    public ApiRespDto<List<LessonStyleTagResponse>> setMyStyleTags(PrincipalUser principal, ArtistStyleSetRequest req) {
         ArtistProfile profile = artistProfileRepository.findByUser_UserId(principal.getUserId())
                 .orElseThrow(()-> new IllegalArgumentException("아티스트 프로필이 없습니다."));
 
         List<Long> ids = (req.styleTagIds() == null) ? List.of() : req.styleTagIds().stream().distinct().toList();
 
+        if (ids.size() > 5) {
+            throw  new IllegalArgumentException("스타일 태그는 최대 5개까지 선택 가능합니다.");
+        }
+
         // 기존 매핑 전부 삭제 후 set
         artistStyleMapRepository.deleteByArtistProfile_ArtistProfileId(profile.getArtistProfileId());
+        artistStyleMapRepository.flush();
 
         if (ids.isEmpty()) {
-            return new ApiRespDto<>("success","아티스트 스타일 태그가 비어있습니다.", null);
+            return new ApiRespDto<>("success","아티스트 스타일 태그가 비어있습니다.", List.of());
         }
 
         // 태그 엔티티 조회
         List<LessonStyleTag> tags = lessonStyleTagRepository.findAllById(ids);
         if (tags.size() != ids.size()){ // 실제 조회댄 태그 개수
-            return new ApiRespDto<>("failed", "존재하지 않는 스타일 태그가 포함되어 있습니다.", null);
+            return new ApiRespDto<>("failed", "존재하지 않는 스타일 태그가 포함되어 있습니다.", List.of());
         }
 
         // 재생성(수정)
@@ -74,7 +79,11 @@ public class ArtistStyleService {
 
         artistStyleMapRepository.saveAll(maps);
 
-        return new ApiRespDto<>("success", "아티스트 스타일 태그가 저장되었습니다.", null);
+        List<LessonStyleTagResponse> resp = tags.stream()
+                .map(t -> new LessonStyleTagResponse(t.getLessonStyleTagId(), t.getStyleName()))
+                .toList();
+
+        return new ApiRespDto<>("success", "아티스트 스타일 태그가 저장되었습니다.", resp);
 
     }
 }
